# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Daily Management Express API",
    "description": "Daily Management REST API documentation for mobile client and server integration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter the Better Auth session token"
      }
    }
  },
  "paths": {
...
    "/api/accounts": {
      "get": {
        "summary": "Get all accounts",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of accounts"
          }
        }
      },
      "post": {
        "summary": "Create an account",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Cash Wallet" },
                  "type": { "type": "string", "enum": ["cash", "debit_card", "bank_account", "credit_card"], "example": "cash" },
                  "balance": { "type": "number", "example": 1000 },
                  "credit_card_details": {
                    "type": "object",
                    "properties": {
                      "total_limit": { "type": "number" },
                      "current_debt": { "type": "number" },
                      "statement_day": { "type": "integer" },
                      "payment_due_day": { "type": "integer" }
                    }
                  }
                },
                "required": ["name", "type", "balance"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully"
          }
        }
      }
    },
    "/api/accounts/{id}": {
      "put": {
        "summary": "Update an account",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Updated Wallet Name" },
                  "balance": { "type": "number", "example": 1200 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Account updated successfully" }
        }
      },
      "delete": {
        "summary": "Delete an account",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
    "/api/categories": {
      "get": {
        "summary": "Get all categories",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of categories" }
        }
      },
      "post": {
        "summary": "Create a category",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Shopping" },
                  "type": { "type": "string", "enum": ["income", "expense"], "example": "expense" },
                  "icon": { "type": "string", "example": "cart" },
                  "color": { "type": "string", "example": "#ef4444" }
                },
                "required": ["name", "type", "icon", "color"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Category created successfully" }
        }
      }
    },
    "/api/categories/{id}": {
      "delete": {
        "summary": "Delete a category",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
...
  }
}